sched, fair: Let minimally loaded cpu balance the group
Currently we let the leftmost (or first idle) cpu acend the
sched_domain tree and perform load-balancing. The result is that the
busiest cpu in the group might be performing this function and pull
more load to itself. The next load balance pass will then try to
equalize this again.

Change this to pick the least loaded cpu to perform higher domain
balancing.

Signed-off-by: Peter Zijlstra <a.p.zijlstra@chello.nl>
modified for Linux 3.0 from Linux 3.4+

commit 9b5780551b070b8d696721e47caad00a8fe3ceb8 1 parent 5fe5cc67b6
faux123 authored 19 days ago drod2169 committed 6 days ago