cpufreq_pegasusq: remove averaging of the average rq
the reason of the additional averaging was to evaluate the runqueue for the total sampling interval instead of single sampling window. However, some peole reported lagginess and I suspect it may be related with this one. so, let's try to disable it and see if it helps.
+ some additional changes to make the governor more responsive.

commit 64cc191a647b5fdd24d14a194b4506188d8c429c 1 parent 04d4ae120c
gokhanmoral authored 16 hours ago