More optimization for NEON flag.
If the selected floating-point hardware includes the NEON extension (e.g. -mfpu=`neon'),
note that floating-point operations are not generated by GCC's auto-vectorization pass unless -funsafe-math-optimizations is also specified.
This is because NEON hardware does not fully implement the IEEE 754 standard for floating-point arithmetic
(in particular denormal values are treated as zero), so the use of NEON instructions may lead to a loss of precision.